+++ src/common/KokkosKernels_SparseUtils.hpp
@@ -51,7 +51,6 @@
 #include "KokkosKernels_ExecSpaceUtils.hpp"
 #include <vector>
 #include "KokkosKernels_PrintUtils.hpp"
-#include "KokkosKernels_Sorting.hpp"
 
 #ifdef KOKKOSKERNELS_HAVE_PARALLEL_GNUSORT
 #include<parallel/algorithm>
@@ -934,453 +933,6 @@ void graph_min_max_degree(const rowmap_t& rowmap, ordinal_t& min_degree, ordinal
       Reducer(result));
   min_degree = result.min_val;
   max_degree = result.max_val;
-}
-
-template<typename execution_space, typename rowmap_t, typename entries_t, typename values_t>
-struct SortCrsMatrixFunctor
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-  using scalar_t = typename values_t::non_const_value_type;
-  using team_mem = typename Kokkos::TeamPolicy<execution_space>::member_type;
-
-  SortCrsMatrixFunctor(bool usingRangePol, const rowmap_t& rowmap_, const entries_t& entries_, const values_t& values_)
-    : rowmap(rowmap_), entries(entries_), values(values_)
-  {
-    if(usingRangePol)
-    {
-      entriesAux = entries_t(Kokkos::ViewAllocateWithoutInitializing("Entries aux"),
-          entries.extent(0));
-      valuesAux = values_t(Kokkos::ViewAllocateWithoutInitializing("Values aux"),
-          values.extent(0));
-    }
-    //otherwise, aux arrays won't be allocated (sorting in place)
-  }
-
-  KOKKOS_INLINE_FUNCTION void operator()(const lno_t i) const
-  {
-    size_type rowStart = rowmap(i);
-    size_type rowEnd = rowmap(i + 1);
-    lno_t rowNum = rowEnd - rowStart;
-    //Radix sort requires unsigned keys for comparison
-    using unsigned_lno_t = typename std::make_unsigned<lno_t>::type;
-    KokkosKernels::Impl::SerialRadixSort2<lno_t, unsigned_lno_t, scalar_t>(
-        (unsigned_lno_t*) entries.data() + rowStart,
-        (unsigned_lno_t*) entriesAux.data() + rowStart,
-        values.data() + rowStart,
-        valuesAux.data() + rowStart, rowNum);
-  }
-
-  KOKKOS_INLINE_FUNCTION void operator()(const team_mem t) const
-  {
-    size_type i = t.league_rank();
-    size_type rowStart = rowmap(i);
-    size_type rowEnd = rowmap(i + 1);
-    lno_t rowNum = rowEnd - rowStart;
-    KokkosKernels::Impl::TeamBitonicSort2<lno_t, lno_t, scalar_t, team_mem>
-      (entries.data() + rowStart, values.data() + rowStart, rowNum, t);
-  }
-
-  rowmap_t rowmap;
-  entries_t entries;
-  entries_t entriesAux;
-  values_t values;
-  values_t valuesAux;
-};
-
-template<typename execution_space, typename rowmap_t, typename entries_t>
-struct SortCrsGraphFunctor
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-  using team_mem = typename Kokkos::TeamPolicy<execution_space>::member_type;
-
-  SortCrsGraphFunctor(bool usingRangePol, const rowmap_t& rowmap_, const entries_t& entries_)
-    : rowmap(rowmap_), entries(entries_)
-  {
-    if(usingRangePol)
-    {
-      entriesAux = entries_t(Kokkos::ViewAllocateWithoutInitializing("Entries aux"),
-          entries.extent(0));
-    }
-    //otherwise, aux arrays won't be allocated (sorting in place)
-  }
-
-  KOKKOS_INLINE_FUNCTION void operator()(const lno_t i) const
-  {
-    size_type rowStart = rowmap(i);
-    size_type rowEnd = rowmap(i + 1);
-    lno_t rowNum = rowEnd - rowStart;
-    //Radix sort requires unsigned keys for comparison
-    using unsigned_lno_t = typename std::make_unsigned<lno_t>::type;
-    KokkosKernels::Impl::SerialRadixSort<lno_t, unsigned_lno_t>(
-        (unsigned_lno_t*) entries.data() + rowStart,
-        (unsigned_lno_t*) entriesAux.data() + rowStart,
-        rowNum);
-  }
-
-  KOKKOS_INLINE_FUNCTION void operator()(const team_mem t) const
-  {
-    size_type i = t.league_rank();
-    size_type rowStart = rowmap(i);
-    size_type rowEnd = rowmap(i + 1);
-    lno_t rowNum = rowEnd - rowStart;
-    KokkosKernels::Impl::TeamBitonicSort<lno_t, lno_t, team_mem>
-      (entries.data() + rowStart, rowNum, t);
-  }
-
-  rowmap_t rowmap;
-  entries_t entries;
-  entries_t entriesAux;
-};
-
-// Sort a CRS matrix: within each row, sort entries ascending by column.
-// At the same time, permute the values.
-template<typename execution_space, typename rowmap_t, typename entries_t, typename values_t>
-void sort_crs_matrix(const rowmap_t& rowmap, const entries_t& entries, const values_t& values)
-{
-  using lno_t = typename entries_t::non_const_value_type;
-  using team_pol = Kokkos::TeamPolicy<execution_space>;
-  bool useRadix = !kk_is_gpu_exec_space<execution_space>();
-  lno_t numRows = rowmap.extent(0) ? rowmap.extent(0) - 1 : 0;
-  if(numRows == 0)
-    return;
-  SortCrsMatrixFunctor<execution_space, rowmap_t, entries_t, values_t>
-    funct(useRadix, rowmap, entries, values);
-  if(useRadix)
-  {
-    Kokkos::parallel_for("sort_crs_matrix", Kokkos::RangePolicy<execution_space>(0, numRows), funct);
-  }
-  else
-  {
-    //Try to get teamsize to be largest power of 2 not greater than avg entries per row
-    //TODO (probably important for performnce): add thread-level sort also, and use that
-    //for small avg degree. But this works for now.
-    lno_t idealTeamSize = 1;
-    lno_t avgDeg = (entries.extent(0) + numRows - 1) / numRows;
-    while(idealTeamSize < avgDeg / 2)
-    {
-      idealTeamSize *= 2;
-    }
-    team_pol temp(numRows, 1);
-    lno_t maxTeamSize = temp.team_size_max(funct, Kokkos::ParallelForTag());
-    lno_t teamSize = std::min(idealTeamSize, maxTeamSize);
-    Kokkos::parallel_for("sort_crs_matrix", team_pol(numRows, teamSize), funct);
-  }
-}
-
-template <typename crsMat_t>
-void sort_crs_matrix(const crsMat_t& A)
-{
-  //Note: rowmap_t has const values, but that's OK as sorting doesn't modify it
-  using rowmap_t = typename crsMat_t::row_map_type;
-  using entries_t = typename crsMat_t::index_type::non_const_type;
-  using values_t = typename crsMat_t::values_type::non_const_type;
-  using exec_space = typename crsMat_t::execution_space;
-  //NOTE: the rowmap of a StaticCrsGraph is const-valued, but the
-  //entries and CrsMatrix values are non-const (so sorting them directly
-  //is allowed)
-  sort_crs_matrix<exec_space, rowmap_t, entries_t, values_t>
-    (A.graph.row_map, A.graph.entries, A.values);
-}
-
-// Sort a CRS graph: within each row, sort entries ascending by column.
-template<typename execution_space, typename rowmap_t, typename entries_t>
-void sort_crs_graph(const rowmap_t& rowmap, const entries_t& entries)
-{
-  using lno_t = typename entries_t::non_const_value_type;
-  using team_pol = Kokkos::TeamPolicy<execution_space>;
-  bool useRadix = !kk_is_gpu_exec_space<execution_space>();
-  lno_t numRows = rowmap.extent(0) ? rowmap.extent(0) - 1 : 0;
-  if(numRows == 0)
-    return;
-  SortCrsGraphFunctor<execution_space, rowmap_t, entries_t>
-    funct(useRadix, rowmap, entries);
-  if(useRadix)
-  {
-    Kokkos::parallel_for("sort_crs_graph", Kokkos::RangePolicy<execution_space>(0, numRows), funct);
-  }
-  else
-  {
-    //Try to get teamsize to be largest power of 2 less than or equal to
-    //half the entries per row. 0.5 * #entries is bitonic's parallelism within a row.
-    //TODO (probably important for performnce): add thread-level sort also, and use that
-    //for small avg degree. But this works for now.
-    lno_t idealTeamSize = 1;
-    lno_t avgDeg = (entries.extent(0) + numRows - 1) / numRows;
-    while(idealTeamSize < avgDeg / 2)
-    {
-      idealTeamSize *= 2;
-    }
-    team_pol temp(numRows, 1);
-    lno_t maxTeamSize = temp.team_size_max(funct, Kokkos::ParallelForTag());
-    lno_t teamSize = std::min(idealTeamSize, maxTeamSize);
-    Kokkos::parallel_for("sort_crs_graph", team_pol(numRows, teamSize), funct);
-  }
-}
-
-template<typename rowmap_t, typename entries_t>
-struct MergedRowmapFunctor
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-  using c_rowmap_t = typename rowmap_t::const_type;
-
-  //Precondition: entries are sorted within each row
-  MergedRowmapFunctor(const rowmap_t& mergedCounts_, const c_rowmap_t& rowmap_, const entries_t& entries_)
-    : mergedCounts(mergedCounts_), rowmap(rowmap_), entries(entries_)
-  {}
-
-  KOKKOS_INLINE_FUNCTION void operator()(lno_t row, size_type& lnewNNZ) const
-  {
-    size_type rowBegin = rowmap(row);
-    size_type rowEnd = rowmap(row + 1);
-    if(rowEnd == rowBegin)
-    {
-      //Row was empty to begin with
-      mergedCounts(row) = 0;
-      return;
-    }
-    //Otherwise, the first entry in the row exists
-    lno_t uniqueEntries = 1;
-    for(size_type j = rowBegin + 1; j < rowEnd; j++)
-    {
-      if(entries(j - 1) != entries(j))
-        uniqueEntries++;
-    }
-    mergedCounts(row) = uniqueEntries;
-    lnewNNZ += uniqueEntries;
-    if(row == lno_t((rowmap.extent(0) - 1) - 1))
-      mergedCounts(row + 1) = 0;
-  }
-
-  rowmap_t mergedCounts;
-  c_rowmap_t rowmap;
-  entries_t entries;
-};
-
-template<typename rowmap_t, typename entries_t, typename values_t>
-struct MatrixMergedEntriesFunctor
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-  using scalar_t = typename values_t::non_const_value_type;
-
-  //Precondition: entries are sorted within each row
-  MatrixMergedEntriesFunctor(
-      const rowmap_t& rowmap_, const entries_t& entries_, const values_t& values_,
-      const rowmap_t& mergedRowmap_, const entries_t& mergedEntries_, const values_t& mergedValues_)
-    : rowmap(rowmap_), entries(entries_), values(values_),
-    mergedRowmap(mergedRowmap_), mergedEntries(mergedEntries_), mergedValues(mergedValues_)
-  {}
-
-  KOKKOS_INLINE_FUNCTION void operator()(lno_t row) const
-  {
-    size_type rowBegin = rowmap(row);
-    size_type rowEnd = rowmap(row + 1);
-    if(rowEnd == rowBegin)
-    {
-      //Row was empty to begin with, nothing to do
-      return;
-    }
-    //Otherwise, accumulate the value for each column
-    scalar_t accumVal = values(rowBegin);
-    lno_t accumCol = entries(rowBegin);
-    size_type insertPos = mergedRowmap(row);
-    for(size_type j = rowBegin + 1; j < rowEnd; j++)
-    {
-      if(accumCol == entries(j))
-      {
-        //accumulate
-        accumVal += values(j);
-      }
-      else
-      {
-        //write out and reset
-        mergedValues(insertPos) = accumVal;
-        mergedEntries(insertPos) = accumCol;
-        insertPos++;
-        accumVal = values(j);
-        accumCol = entries(j);
-      }
-    }
-    //always left with the last unique entry
-    mergedValues(insertPos) = accumVal;
-    mergedEntries(insertPos) = accumCol;
-  }
-
-  rowmap_t rowmap;
-  entries_t entries;
-  values_t values;
-  rowmap_t mergedRowmap;
-  entries_t mergedEntries;
-  values_t mergedValues;
-};
-
-template<typename rowmap_t, typename entries_t>
-struct GraphMergedEntriesFunctor
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-
-  //Precondition: entries are sorted within each row
-  GraphMergedEntriesFunctor(
-      const rowmap_t& rowmap_, const entries_t& entries_,
-      const rowmap_t& mergedRowmap_, const entries_t& mergedEntries_)
-    : rowmap(rowmap_), entries(entries_),
-    mergedRowmap(mergedRowmap_), mergedEntries(mergedEntries_)
-  {}
-
-  KOKKOS_INLINE_FUNCTION void operator()(lno_t row) const
-  {
-    size_type rowBegin = rowmap(row);
-    size_type rowEnd = rowmap(row + 1);
-    if(rowEnd == rowBegin)
-    {
-      //Row was empty to begin with, nothing to do
-      return;
-    }
-    //Otherwise, accumulate the value for each column
-    lno_t accumCol = entries(rowBegin);
-    size_type insertPos = mergedRowmap(row);
-    for(size_type j = rowBegin + 1; j < rowEnd; j++)
-    {
-      if(accumCol != entries(j))
-      {
-        //write out and reset
-        mergedEntries(insertPos) = accumCol;
-        insertPos++;
-        accumCol = entries(j);
-      }
-    }
-    //always left with the last unique entry
-    mergedEntries(insertPos) = accumCol;
-  }
-
-  rowmap_t rowmap;
-  entries_t entries;
-  rowmap_t mergedRowmap;
-  entries_t mergedEntries;
-};
-
-//Sort the rows of matrix, and merge duplicate entries.
-template<typename crsMat_t>
-crsMat_t sort_and_merge_matrix(const crsMat_t& A)
-{
-  using c_rowmap_t = typename crsMat_t::row_map_type;
-  using rowmap_t = typename crsMat_t::row_map_type::non_const_type;
-  using entries_t = typename crsMat_t::index_type::non_const_type;
-  using values_t = typename crsMat_t::values_type::non_const_type;
-  using size_type = typename rowmap_t::non_const_value_type;
-  using exec_space = typename crsMat_t::execution_space;
-  using range_t = Kokkos::RangePolicy<exec_space>;
-  sort_crs_matrix(A);
-  //Count entries per row into a new rowmap, in terms of merges that can be done
-  rowmap_t mergedRowmap(Kokkos::ViewAllocateWithoutInitializing("SortedMerged rowmap"), A.numRows() + 1);
-  size_type numCompressedEntries = 0;
-  Kokkos::parallel_reduce(range_t(0, A.numRows()),
-      MergedRowmapFunctor<rowmap_t, entries_t>(mergedRowmap, A.graph.row_map, A.graph.entries), numCompressedEntries);
-  //Prefix sum to get rowmap
-  kk_exclusive_parallel_prefix_sum<rowmap_t, exec_space>(A.numRows() + 1, mergedRowmap);
-  entries_t mergedEntries("SortedMerged entries", numCompressedEntries);
-  values_t mergedValues("SortedMerged values", numCompressedEntries);
-  //Compute merged entries and values
-  Kokkos::parallel_for(range_t(0, A.numRows()),
-      MatrixMergedEntriesFunctor<c_rowmap_t, entries_t, values_t>
-      (A.graph.row_map, A.graph.entries, A.values,
-       mergedRowmap, mergedEntries, mergedValues));
-  //Finally, construct the new compressed matrix
-  return crsMat_t("SortedMerged", A.numRows(), A.numCols(), numCompressedEntries,
-      mergedValues, mergedRowmap, mergedEntries);
-}
-
-template<typename exec_space, typename rowmap_t, typename entries_t>
-void sort_and_merge_graph(
-    const typename rowmap_t::const_type& rowmap_in, const entries_t& entries_in,
-    rowmap_t& rowmap_out, entries_t& entries_out)
-{
-  using size_type = typename rowmap_t::non_const_value_type;
-  using lno_t = typename entries_t::non_const_value_type;
-  using range_t = Kokkos::RangePolicy<exec_space>;
-  using const_rowmap_t = typename rowmap_t::const_type;
-  lno_t numRows = rowmap_in.extent(0);
-  if(numRows <= 1)
-  {
-    //Matrix has zero rows
-    rowmap_out = rowmap_t();
-    entries_out = entries_t();
-    return;
-  }
-  numRows--;
-  //Sort in place
-  sort_crs_graph<exec_space, const_rowmap_t, entries_t>(rowmap_in, entries_in);
-  //Count entries per row into a new rowmap, in terms of merges that can be done
-  rowmap_out = rowmap_t(Kokkos::ViewAllocateWithoutInitializing("SortedMerged rowmap"), numRows + 1);
-  size_type numCompressedEntries = 0;
-  Kokkos::parallel_reduce(range_t(0, numRows),
-      MergedRowmapFunctor<rowmap_t, entries_t>(rowmap_out, rowmap_in, entries_in), numCompressedEntries);
-  //Prefix sum to get rowmap
-  kk_exclusive_parallel_prefix_sum<rowmap_t, exec_space>(numRows + 1, rowmap_out);
-  entries_out = entries_t("SortedMerged entries", numCompressedEntries);
-  //Compute merged entries and values
-  Kokkos::parallel_for(range_t(0, numRows),
-      GraphMergedEntriesFunctor<const_rowmap_t, entries_t>
-      (rowmap_in, entries_in,
-       rowmap_out, entries_out));
-}
-
-template <typename lno_view_t,
-          typename lno_nnz_view_t,
-          typename scalar_view_t,
-
-          typename out_nnz_view_t,
-          typename out_scalar_view_t,
-          typename MyExecSpace>
-void kk_sort_graph(
-    lno_view_t in_xadj,
-    lno_nnz_view_t in_adj,
-    scalar_view_t in_vals,
-
-    out_nnz_view_t out_adj,
-    out_scalar_view_t out_vals){
-  // TODO BMK: can this function be deprecated?
-  typename lno_view_t::HostMirror hr = Kokkos::create_mirror_view (in_xadj);
-  Kokkos::deep_copy (hr, in_xadj);
-  typename lno_nnz_view_t::HostMirror he = Kokkos::create_mirror_view (in_adj);
-  Kokkos::deep_copy (he, in_adj);
-  typename scalar_view_t::HostMirror hv = Kokkos::create_mirror_view (in_vals);
-  Kokkos::deep_copy (hv, in_vals);
-  MyExecSpace().fence();
-
-  typename lno_nnz_view_t::HostMirror heo = Kokkos::create_mirror_view (out_adj);
-  typename scalar_view_t::HostMirror hvo = Kokkos::create_mirror_view (out_vals);
-
-  typedef typename lno_view_t::non_const_value_type size_type;
-  typedef typename lno_nnz_view_t::non_const_value_type lno_t;
-  typedef typename scalar_view_t::non_const_value_type scalar_t;
-
-  lno_t nrows = in_xadj.extent(0) - 1;
-  std::vector <Edge<lno_t, scalar_t> > edges(in_adj.extent(0));
-
-  size_type row_size = 0;
-  for (lno_t i = 0; i < nrows; ++i){
-    for (size_type j = hr(i); j < hr(i + 1); ++j){
-      edges[row_size].src = i;
-      edges[row_size].dst = he(j);
-      edges[row_size++].ew = hv(j);
-    }
-  }
-  std::sort (edges.begin(), edges.begin() + row_size);
-  size_type ne = in_adj.extent(0);
-  for(size_type i = 0; i < ne; ++i){
-    heo(i) = edges[i].dst;
-    hvo(i) = edges[i].ew;
-  }
-
-
-  Kokkos::deep_copy (out_adj, heo);
-  Kokkos::deep_copy (out_vals, hvo);
-  MyExecSpace().fence();
 }
 
 /*
@@ -2341,94 +1893,6 @@ void kk_create_incidence_tranpose_matrix_from_lower_triangle(
     });
   }
 
-template <typename row_map_view_t,
-          typename cols_view_t,
-          typename out_row_map_view_t,
-          typename out_cols_view_t,
-          typename exec_space
-          >
-void kk_create_incidence_matrix_from_lower_triangle(
-    typename cols_view_t::non_const_value_type nr,
-    row_map_view_t in_lower_rowmap,
-    cols_view_t in_lower_entries,
-    out_row_map_view_t &out_rowmap,
-    out_cols_view_t &out_entries,
-    bool use_dynamic_scheduling = false,
-    bool chunksize = 4){
-
-  //typedef typename row_map_view_t::const_type const_row_map_view_t;
-  //typedef typename cols_view_t::const_type   const_cols_view_t;
-
-  typedef typename row_map_view_t::non_const_value_type size_type;
-  typedef typename cols_view_t::non_const_value_type lno_t;
-
-
-
-  //const size_type *rowmap = in_rowmap.data();
-  //const lno_t *entries= in_entries.data();
-  const size_type ne = in_lower_entries.extent(0);
-  typedef Kokkos::RangePolicy<exec_space> my_exec_space;
-  out_rowmap = out_row_map_view_t("LL", nr+1);
-
-  Kokkos::parallel_for(my_exec_space(0, ne),
-      KOKKOS_LAMBDA(const lno_t& i) {
-    typedef typename std::remove_reference< decltype( out_rowmap[0] ) >::type atomic_incr_type;
-    Kokkos::atomic_fetch_add(&(out_rowmap[in_lower_entries[i]]), atomic_incr_type(1));
-  });
-
-  exec_space().fence();
-  kk_exclusive_parallel_prefix_sum<out_row_map_view_t, exec_space>(nr+1, out_rowmap);
-
-  exec_space().fence();
-  Kokkos::parallel_for("KokkosKernels::Common::CreateIncidenceTransposeMatrixFromLowerTriangle::S0", my_exec_space(0, nr + 1),
-      KOKKOS_LAMBDA(const lno_t& i) {
-    out_rowmap[i] += in_lower_rowmap[i];
-  });
-
-  out_row_map_view_t out_rowmap_copy (Kokkos::ViewAllocateWithoutInitializing("tmp"), nr+1);
-  Kokkos::deep_copy(out_rowmap_copy, out_rowmap);
-
-  out_entries = out_cols_view_t(Kokkos::ViewAllocateWithoutInitializing("LL"), 2*ne);
-
-  Kokkos::parallel_for("KokkosKernels::Common::CreateIncidenceTransposeMatrixFromLowerTriangle::S1", my_exec_space(0, nr),
-      KOKKOS_LAMBDA(const size_type& row) {
-    size_type begin = in_lower_rowmap(row);
-    lno_t row_size = in_lower_rowmap(row + 1) - begin;
-
-
-
-    size_type out_begin = out_rowmap_copy(row);
-    out_rowmap_copy(row)  += row_size;
-    //lno_t row_size = out_rowmap_copy(row + 1) - begin;
-
-    for (int i = 0; i < row_size; ++i){
-      size_type edge_ind = i + begin;
-      out_entries[out_begin + i] = edge_ind;
-    }
-  });
-  exec_space().fence();
-  Kokkos::parallel_for("KokkosKernels::Common::CreateIncidenceTransposeMatrixFromLowerTriangle::S2", my_exec_space(0, ne),
-      KOKKOS_LAMBDA(const size_type& edge_ind) {
-    lno_t col = in_lower_entries[edge_ind];
-    typedef typename std::remove_reference< decltype( out_rowmap_copy(0) ) >::type atomic_incr_type;
-    size_type write_ind = Kokkos::atomic_fetch_add(&(out_rowmap_copy(col)), atomic_incr_type(1));
-    out_entries[write_ind] = edge_ind;
-  });
-
-  out_cols_view_t tmp ("kk", ne * 2);
-  out_cols_view_t outcols ("kk", ne * 2);
-
-  kk_sort_graph<out_row_map_view_t, out_cols_view_t, out_cols_view_t, out_cols_view_t, out_cols_view_t, exec_space>
-      (out_rowmap, out_entries,
-          tmp,
-
-          outcols,
-      tmp);
-
-      out_entries = outcols;
-}
-
-
 
 template <typename row_map_view_t,
           typename cols_view_t,
@@ -2551,22 +2015,7 @@ void kk_create_incidence_matrix_from_original_matrix(
       KOKKOS_LAMBDA(const lno_t& i) {
     out_rowmap[i] = in_rowmap[i];
   });
-
-
-  /*
-
-  out_cols_view_t tmp ("kk", ne * 2);
-  out_cols_view_t outcols ("kk", ne * 2);
-
-  kk_sort_graph<out_row_map_view_t, out_cols_view_t, out_cols_view_t, out_cols_view_t, out_cols_view_t, exec_space>
-      (out_rowmap, out_entries,
-          tmp,
-
-          outcols,
-      tmp);
-
-      out_entries = outcols;*/
-  }
+}
