+++ src/sparse/impl/KokkosSparse_twostage_gauss_seidel_impl.hpp
@@ -845,9 +845,9 @@ namespace KokkosSparse{
           if (sptrsv_algo == SPTRSVAlgorithm::SPTRSV_CUSPARSE) { // symbolic with CuSparse needs values
             // CuSparse needs matrix sorted by column indexes for each row
             // TODO: may need to move this to symbolic/numeric of sptrsv
-            KokkosKernels::Impl::sort_crs_matrix <execution_space, const_row_map_view_t, entries_view_t, values_view_t>
+            KokkosKernels::sort_crs_matrix <execution_space, const_row_map_view_t, entries_view_t, values_view_t>
               (rowmap_viewL, column_viewL, values_viewL);
-            KokkosKernels::Impl::sort_crs_matrix <execution_space, const_row_map_view_t, entries_view_t, values_view_t>
+            KokkosKernels::sort_crs_matrix <execution_space, const_row_map_view_t, entries_view_t, values_view_t>
               (rowmap_viewU, column_viewU, values_viewU);
 
             // now do symbolic
