+++ unit_test/sparse/Test_Sparse_spgemm.hpp
@@ -47,6 +47,7 @@
 #include <Kokkos_Core.hpp>
 
 #include "KokkosKernels_SparseUtils.hpp"
+#include "KokkosKernels_Sorting.hpp"
 #include <Kokkos_Concepts.hpp>
 #include <string>
 #include <stdexcept>
@@ -202,27 +203,6 @@ bool is_same_matrix(crsMat_t output_mat_actual, crsMat_t output_mat_reference){
   size_t nentries_reference = output_mat_reference.graph.entries.extent(0) ;
   size_t nvals_reference = output_mat_reference.values.extent(0);
 
-
-  lno_nnz_view_t h_ent_actual (Kokkos::ViewAllocateWithoutInitializing("h_ent_actual"), nentries_actual);
-  scalar_view_t h_vals_actual (Kokkos::ViewAllocateWithoutInitializing("h_vals_actual"), nvals_actual);
-
-
-  KokkosKernels::Impl::kk_sort_graph<typename graph_t::row_map_type,
-    typename graph_t::entries_type,
-    typename crsMat_t::values_type,
-    lno_nnz_view_t,
-    scalar_view_t,
-    typename device::execution_space
-    >(
-    output_mat_actual.graph.row_map, 
-    output_mat_actual.graph.entries, 
-    output_mat_actual.values,
-    h_ent_actual, h_vals_actual
-  );
-
-  lno_nnz_view_t h_ent_reference (Kokkos::ViewAllocateWithoutInitializing("h_ent_reference"), nentries_reference);
-  scalar_view_t h_vals_reference (Kokkos::ViewAllocateWithoutInitializing("h_vals_reference"), nvals_reference);
-
   if (nrows_actual != nrows_reference) { 
      std::cout << "nrows_actual:" << nrows_actual << " nrows_reference:" << nrows_reference << std::endl;
      return false;
@@ -236,19 +216,8 @@ bool is_same_matrix(crsMat_t output_mat_actual, crsMat_t output_mat_reference){
     return false;
   }
 
-  KokkosKernels::Impl::kk_sort_graph
-      <typename graph_t::row_map_type,
-      typename graph_t::entries_type,
-      typename crsMat_t::values_type,
-      lno_nnz_view_t,
-      scalar_view_t,
-      typename device::execution_space
-      >(
-      output_mat_reference.graph.row_map, 
-      output_mat_reference.graph.entries, 
-      output_mat_reference.values,
-      h_ent_reference, h_vals_reference
-    );
+  KokkosKernels::sort_crs_matrix(output_mat_actual);
+  KokkosKernels::sort_crs_matrix(output_mat_reference);
 
   bool is_identical = true;
   is_identical = KokkosKernels::Impl::kk_is_identical_view
@@ -266,12 +235,12 @@ bool is_same_matrix(crsMat_t output_mat_actual, crsMat_t output_mat_reference){
 
   is_identical = KokkosKernels::Impl::kk_is_identical_view
       <lno_nnz_view_t, lno_nnz_view_t, typename lno_nnz_view_t::value_type,
-      typename device::execution_space>(h_ent_actual, h_ent_reference, 0 );
+      typename device::execution_space>(output_mat_actual.graph.entries, output_mat_reference.graph.entries, 0 );
 
   if (!is_identical) {
     std::cout << "entries are different." << std::endl;
-    KokkosKernels::Impl::kk_print_1Dview(h_ent_actual);
-    KokkosKernels::Impl::kk_print_1Dview(h_ent_reference);
+    KokkosKernels::Impl::kk_print_1Dview(output_mat_actual.graph.entries);
+    KokkosKernels::Impl::kk_print_1Dview(output_mat_reference.graph.entries);
     return false;
   }
 
@@ -282,7 +251,7 @@ bool is_same_matrix(crsMat_t output_mat_actual, crsMat_t output_mat_reference){
 
   is_identical = KokkosKernels::Impl::kk_is_relatively_identical_view
       <scalar_view_t, scalar_view_t, eps_type,
-      typename device::execution_space>(h_vals_actual, h_vals_reference, eps);
+      typename device::execution_space>(output_mat_actual.values, output_mat_reference.values, eps);
 
   if (!is_identical) {
     std::cout << "values are different." << std::endl;
